+++ pl_examples/test_examples.py
@@ -25,20 +25,6 @@ def test_gpu_template(cli_args):
         run_cli()
 
 
-@pytest.mark.parametrize(
-    'cli_args',
-    ['--max_epochs 1 --max_steps 3 --num_nodes 1 --gpus 2 --dist_backend ddp_spawn --bn_sync']
-)
-@pytest.mark.skipif(torch.cuda.device_count() < 2, reason="test requires multi-GPU machine")
-def test_sync_bn(cli_args):
-    """Test running CLI for an example with sync bn."""
-    from pl_examples.basic_examples.sync_bn import run_cli
-
-    cli_args = cli_args.split(' ') if cli_args else []
-    with mock.patch("argparse._sys.argv", ["any.py"] + cli_args):
-        run_cli()
-
-
 # @pytest.mark.parametrize('cli_args', ['--max_epochs 1 --max_steps 3 --num_nodes 1 --gpus 2'])
 # @pytest.mark.skipif(torch.cuda.device_count() < 2, reason="test requires multi-GPU machine")
 # def test_multi_node_ddp(cli_args):
